updated profile page structure
