+++ schedulite.clientapp/src/screens/ScheduleSelectionScreen/Block.css
