+++ schedulite.clientapp/src/screens/ScheduleSelectionScreen/BlockPage.tsx
